refactor(auth): remove unduplicate on auth guard (web)
